[STREAMPIPES-563] Started to restructured code of data lake sink
